planning: naming changes for traffic-light scenario
