+++ modules/planning/conf/scenario/traffic_light_unprotected_right_turn_config.pb.txt
