+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/BUILD
@@ -3,18 +3,18 @@ load("//tools:cpplint.bzl", "cpplint")
 package(default_visibility = ["//visibility:public"])
 
 cc_library(
-    name = "traffic_light_right_turn_unprotected_scenario",
+    name = "traffic_light_unprotected_right_turn_scenario",
     srcs = [
         "stage_creep.cc",
         "stage_intersection_cruise.cc",
         "stage_stop.cc",
-        "traffic_light_right_turn_unprotected_scenario.cc",
+        "traffic_light_unprotected_right_turn_scenario.cc",
     ],
     hdrs = [
         "stage_creep.h",
         "stage_intersection_cruise.h",
         "stage_stop.h",
-        "traffic_light_right_turn_unprotected_scenario.h",
+        "traffic_light_unprotected_right_turn_scenario.h",
     ],
     copts = ["-DMODULE_NAME=\\\"planning\\\""],
     deps = [
@@ -33,10 +33,10 @@ cc_library(
 cc_test(
     name = "traffic_light_right_turn_unprotected_scenario_test",
     size = "small",
-    srcs = ["traffic_light_right_turn_unprotected_scenario_test.cc"],
+    srcs = ["traffic_light_unprotected_right_turn_scenario_test.cc"],
     data = ["//modules/planning:planning_conf"],
     deps = [
-        ":traffic_light_right_turn_unprotected_scenario",
+        ":traffic_light_unprotected_right_turn_scenario",
         "@gtest//:main",
     ],
 )
