+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_creep.cc
@@ -18,9 +18,10 @@
  * @file
  **/
 
+#include <string>
 #include <vector>
 
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/stage_stop.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_creep.h"
 
 #include "modules/perception/proto/perception_obstacle.pb.h"
 #include "modules/perception/proto/traffic_light_detection.pb.h"
@@ -37,75 +38,73 @@ namespace planning {
 namespace scenario {
 namespace traffic_light {
 
-using common::TrajectoryPoint;
 using common::time::Clock;
+using common::TrajectoryPoint;
 using hdmap::PathOverlap;
 using perception::TrafficLight;
 
-Stage::StageStatus StageStop::Process(
+Stage::StageStatus StageCreep::Process(
     const TrajectoryPoint& planning_init_point, Frame* frame) {
-  ADEBUG << "stage: Stop";
+  ADEBUG << "stage: Creep";
   CHECK_NOTNULL(frame);
 
   scenario_config_.CopyFrom(GetContext()->scenario_config);
 
+  if (!config_.enabled()) {
+    return FinishStage();
+  }
+
   bool plan_ok = ExecuteTaskOnReferenceLine(planning_init_point, frame);
   if (!plan_ok) {
-    AERROR << "TrafficLightRightTurnUnprotectedStop planning error";
+    AERROR << "StageCreep planning error";
   }
 
   const auto& reference_line_info = frame->reference_line_info().front();
 
-  // check if the traffic_light is still along reference_line
-  std::string traffic_light_id = GetContext()->traffic_light_id;
+  // check if the traffic_light is still along referenceline
+  std::string traffic_light_overlap_id = GetContext()->traffic_light_id;
   const std::vector<PathOverlap>& traffic_light_overlaps =
       reference_line_info.reference_line().map_path().signal_overlaps();
   auto traffic_light_overlap_it =
       std::find_if(traffic_light_overlaps.begin(), traffic_light_overlaps.end(),
-                   [&traffic_light_id](const PathOverlap& overlap) {
-                     return overlap.object_id == traffic_light_id;
+                   [&traffic_light_overlap_id](const PathOverlap& overlap) {
+                     return overlap.object_id == traffic_light_overlap_id;
                    });
   if (traffic_light_overlap_it == traffic_light_overlaps.end()) {
     next_stage_ = ScenarioConfig::NO_STAGE;
     return Stage::FINISHED;
   }
 
-  const double adc_front_edge_s = reference_line_info.AdcSlBoundary().end_s();
-  const double distance_adc_pass_traffic_light =
-      adc_front_edge_s - traffic_light_overlap_it->start_s;
-  constexpr double kPassStopLineBuffer = 1.0;  // unit: m
-
-  // passed stop line too far
-  if (distance_adc_pass_traffic_light > kPassStopLineBuffer) {
-    return FinishStage();
-  }
-
-  // check on wait-time
-  auto start_time = GetContext()->stop_start_time;
-  const double wait_time = Clock::NowInSeconds() - start_time;
-  ADEBUG << "stop_start_time[" << start_time << "] wait_time[" << wait_time
-         << "]";
-  if (wait_time > scenario_config_.red_light_right_turn_stop_duration()) {
-    return FinishStage();
-  }
-
   // check on traffic light color
   if (PlanningContext::GetScenarioInfo()->traffic_light_color ==
       TrafficLight::GREEN) {
-    next_stage_ = ScenarioConfig::
-        TRAFFIC_LIGHT_UNPROTECTED_RIGHT_TURN_INTERSECTION_CRUISE;
-    return Stage::FINISHED;
+    return FinishStage();
   }
 
+  const double wait_time =
+      Clock::NowInSeconds() - GetContext()->creep_start_time;
+  const double timeout = scenario_config_.creep_timeout();
+  if (dynamic_cast<DeciderCreep*>(FindTask(TaskConfig::DECIDER_CREEP))
+          ->CheckCreepDone(*frame, reference_line_info,
+                           traffic_light_overlap_it->end_s,
+                           wait_time, timeout)) {
+    return FinishStage();
+  }
+
+  // set param for PROCEED_WITH_CAUTION_SPEED
+  dynamic_cast<DeciderCreep*>(FindTask(TaskConfig::DECIDER_CREEP))
+      ->SetProceedWithCautionSpeedParam(*frame, reference_line_info,
+                                        traffic_light_overlap_it->end_s);
+
+  plan_ok = ExecuteTaskOnReferenceLine(planning_init_point, frame);
+  if (!plan_ok) {
+    AERROR << "StageCreep planning error";
+  }
   return Stage::RUNNING;
 }
 
-Stage::StageStatus StageStop::FinishStage() {
-  PlanningContext::GetScenarioInfo()->stop_done_overlap_id =
-      GetContext()->traffic_light_id;
-  GetContext()->creep_start_time = Clock::NowInSeconds();
-
-  next_stage_ = ScenarioConfig::TRAFFIC_LIGHT_UNPROTECTED_RIGHT_TURN_CREEP;
+Stage::StageStatus StageCreep::FinishStage() {
+  next_stage_ = ScenarioConfig::STOP_SIGN_UNPROTECTED_INTERSECTION_CRUISE;
   return Stage::FINISHED;
 }
