+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_creep.h
@@ -23,7 +23,7 @@
 #include "modules/planning/proto/planning_config.pb.h"
 
 #include "modules/planning/scenarios/stage.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/traffic_light_right_turn_unprotected_scenario.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/traffic_light_unprotected_right_turn_scenario.h"
 
 namespace apollo {
 namespace planning {
@@ -32,17 +32,17 @@ namespace traffic_light {
 
 struct TrafficLightRightTurnUnprotectedContext;
 
-class StageIntersectionCruise : public Stage {
+class StageCreep : public Stage {
  public:
-  explicit StageIntersectionCruise(const ScenarioConfig::StageConfig& config)
-      : Stage(config) {}
+  explicit StageCreep(
+      const ScenarioConfig::StageConfig& config) : Stage(config) {}
 
  private:
   Stage::StageStatus Process(const common::TrajectoryPoint& planning_init_point,
                              Frame* frame) override;
 
   TrafficLightRightTurnUnprotectedContext* GetContext() {
-    return GetContextAs<TrafficLightRightTurnUnprotectedContext>();
+    return Stage::GetContextAs<TrafficLightRightTurnUnprotectedContext>();
   }
 
  private:
