+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_intersection_cruise.cc
@@ -21,10 +21,9 @@
 #include <string>
 #include <vector>
 
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/stage_creep.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_intersection_cruise.h"
 
 #include "modules/perception/proto/perception_obstacle.pb.h"
-#include "modules/perception/proto/traffic_light_detection.pb.h"
 
 #include "cyber/common/log.h"
 #include "modules/common/time/time.h"
@@ -38,25 +37,17 @@ namespace planning {
 namespace scenario {
 namespace traffic_light {
 
-using common::time::Clock;
 using common::TrajectoryPoint;
 using hdmap::PathOverlap;
-using perception::TrafficLight;
 
-Stage::StageStatus StageCreep::Process(
+Stage::StageStatus StageIntersectionCruise::Process(
     const TrajectoryPoint& planning_init_point, Frame* frame) {
-  ADEBUG << "stage: Creep";
+  ADEBUG << "stage: IntersectionCruise";
   CHECK_NOTNULL(frame);
-
-  scenario_config_.CopyFrom(GetContext()->scenario_config);
-
-  if (!config_.enabled()) {
-    return FinishStage();
-  }
 
   bool plan_ok = ExecuteTaskOnReferenceLine(planning_init_point, frame);
   if (!plan_ok) {
-    AERROR << "StageCreep planning error";
+    AERROR << "StageIntersectionCruise plan error";
   }
 
   const auto& reference_line_info = frame->reference_line_info().front();
@@ -71,40 +62,22 @@ Stage::StageStatus StageCreep::Process(
                      return overlap.object_id == traffic_light_overlap_id;
                    });
   if (traffic_light_overlap_it == traffic_light_overlaps.end()) {
-    next_stage_ = ScenarioConfig::NO_STAGE;
-    return Stage::FINISHED;
-  }
-
-  // check on traffic light color
-  if (PlanningContext::GetScenarioInfo()->traffic_light_color ==
-      TrafficLight::GREEN) {
     return FinishStage();
   }
 
-  const double wait_time =
-      Clock::NowInSeconds() - GetContext()->creep_start_time;
-  const double timeout = scenario_config_.creep_timeout();
-  if (dynamic_cast<DeciderCreep*>(FindTask(TaskConfig::DECIDER_CREEP))
-          ->CheckCreepDone(*frame, reference_line_info,
-                           traffic_light_overlap_it->end_s,
-                           wait_time, timeout)) {
+  // check pass intersection
+  // TODO(all): update when pnc-junction is ready
+  constexpr double kIntersectionLength = 10.0;  // unit: m
+  const double adc_back_edge_s = reference_line_info.AdcSlBoundary().start_s();
+  if (adc_back_edge_s - traffic_light_overlap_it->end_s > kIntersectionLength) {
     return FinishStage();
   }
 
-  // set param for PROCEED_WITH_CAUTION_SPEED
-  dynamic_cast<DeciderCreep*>(FindTask(TaskConfig::DECIDER_CREEP))
-      ->SetProceedWithCautionSpeedParam(*frame, reference_line_info,
-                                        traffic_light_overlap_it->end_s);
-
-  plan_ok = ExecuteTaskOnReferenceLine(planning_init_point, frame);
-  if (!plan_ok) {
-    AERROR << "StageCreep planning error";
-  }
   return Stage::RUNNING;
 }
 
-Stage::StageStatus StageCreep::FinishStage() {
-  next_stage_ = ScenarioConfig::STOP_SIGN_UNPROTECTED_INTERSECTION_CRUISE;
+Stage::StageStatus StageIntersectionCruise::FinishStage() {
+  next_stage_ = ScenarioConfig::NO_STAGE;
   return Stage::FINISHED;
 }
