+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_intersection_cruise.h
@@ -20,12 +20,10 @@
 
 #pragma once
 
-#include <string>
-
 #include "modules/planning/proto/planning_config.pb.h"
 
 #include "modules/planning/scenarios/stage.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/traffic_light_right_turn_unprotected_scenario.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/traffic_light_unprotected_right_turn_scenario.h"
 
 namespace apollo {
 namespace planning {
@@ -34,15 +32,15 @@ namespace traffic_light {
 
 struct TrafficLightRightTurnUnprotectedContext;
 
-class StageStop : public Stage {
+class StageIntersectionCruise : public Stage {
  public:
-  explicit StageStop(
-      const ScenarioConfig::StageConfig& config) : Stage(config) {}
+  explicit StageIntersectionCruise(const ScenarioConfig::StageConfig& config)
+      : Stage(config) {}
 
  private:
-  Stage::StageStatus Process(
-      const common::TrajectoryPoint& planning_init_point,
-      Frame* frame) override;
+  Stage::StageStatus Process(const common::TrajectoryPoint& planning_init_point,
+                             Frame* frame) override;
+
   TrafficLightRightTurnUnprotectedContext* GetContext() {
     return GetContextAs<TrafficLightRightTurnUnprotectedContext>();
   }
