+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_stop.h
@@ -20,10 +20,12 @@
 
 #pragma once
 
+#include <string>
+
 #include "modules/planning/proto/planning_config.pb.h"
 
 #include "modules/planning/scenarios/stage.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/traffic_light_right_turn_unprotected_scenario.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/traffic_light_unprotected_right_turn_scenario.h"
 
 namespace apollo {
 namespace planning {
@@ -32,17 +34,17 @@ namespace traffic_light {
 
 struct TrafficLightRightTurnUnprotectedContext;
 
-class StageCreep : public Stage {
+class StageStop : public Stage {
  public:
-  explicit StageCreep(
+  explicit StageStop(
       const ScenarioConfig::StageConfig& config) : Stage(config) {}
 
  private:
-  Stage::StageStatus Process(const common::TrajectoryPoint& planning_init_point,
-                             Frame* frame) override;
-
+  Stage::StageStatus Process(
+      const common::TrajectoryPoint& planning_init_point,
+      Frame* frame) override;
   TrafficLightRightTurnUnprotectedContext* GetContext() {
-    return Stage::GetContextAs<TrafficLightRightTurnUnprotectedContext>();
+    return GetContextAs<TrafficLightRightTurnUnprotectedContext>();
   }
 
  private:
