+++ modules/planning/scenarios/traffic_light/unprotected_right_turn/traffic_light_unprotected_right_turn_scenario.cc
@@ -18,7 +18,7 @@
  * @file
  **/
 
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/traffic_light_right_turn_unprotected_scenario.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/traffic_light_unprotected_right_turn_scenario.h"
 
 #include "modules/perception/proto/perception_obstacle.pb.h"
 #include "modules/perception/proto/traffic_light_detection.pb.h"
@@ -29,9 +29,9 @@
 #include "modules/common/vehicle_state/vehicle_state_provider.h"
 #include "modules/planning/common/frame.h"
 #include "modules/planning/common/planning_context.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/stage_creep.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/stage_intersection_cruise.h"
-#include "modules/planning/scenarios/traffic_light/right_turn_unprotected/stage_stop.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_creep.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_intersection_cruise.h"
+#include "modules/planning/scenarios/traffic_light/unprotected_right_turn/stage_stop.h"
 
 namespace apollo {
 namespace planning {
